src/requirements.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
confusable-homoglyphs==3.3.1
crispy-bootstrap5==2024.10
cryptography==44.0.1
Django==4.2.20
django-admin-honeypot-updated-2021==1.2.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-ipware==7.0.1
django-redis==5.4.0
django-registration==5.1.0
django-taggit==6.1.0
exceptiongroup==1.2.2
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
idna==3.10
outcome==1.3.0.post0
playwright==1.49.1
psycopg==3.2.5
psycopg-binary==3.2.5
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
PySocks==1.7.1
python-decouple==3.8
python-ipware==3.0.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tldextract==5.1.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
